videoflix_backend/requirements.txt
arrow==1.3.0
asgiref==3.8.1
click==8.1.8
cmake==3.31.4
colorama==0.4.6
coverage==7.6.10
diff-match-patch==20241021
Django==5.1.2
django-cors-headers==4.4.0
django-debug-toolbar==5.0.1
django-filter==24.3
django-import-export==4.3.4
django-redis==5.4.0
django-rq==3.0.0
djangorestframework==3.15.2
factory_boy==3.3.1
Faker==35.2.0
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
ninja==1.11.1.3
packaging==24.2
pinocchio==0.4.3
pluggy==1.5.0
psycopg2-binary==2.9.10
Pygments==2.19.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
redis==3.5.2
rq==2.1.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.1
tablib==3.8.0
times==0.7
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
